Move all cli code into a module; hide private impls in private submodule
